Added a parser to generate a syntax tree and interpreter and repl
Parser generates a tree based on the original CFG grammer. Since
the tokenizing process already removes left recursion, it should be safe
to use the original grammer. The syntax tree gets traversed by the interpreter
which replaces ids with values in a passed in look up table of variables.
However there is a bug where the left hand side of an assignment can
have it's id replaced with a int value and then be added to the look up table.
I can make a hack to bypass id value replacement on left hand side of assignment
but I want to see if there's a more elegant way to not do left hand side value
replacement and also do left hand side value updating for new assignments.