Added syntax error and unintialized variables detection
Lexer now raises a syntax error when it's unable to tokenize
the statement. The parser raises a name error when a variable
has not been defined/intialized. The syntax error triggers the repl
to print out a syntax error has occured and the statement is discarded
The name error triggers the repl to print out the variable that
was not defined/intialized. Variables now properly gets replaced
with it's defined value and variables now support redefining so
statements like x=1; and x=x+1; should redefine the x to 2;