Factored out the Number class
I thought I had to wrap everything in the Number class incase
I need to cast a number back into int, but i realized there's no
division so any arithmetic in python won't change the number to a
float. The initial cast of int in the lexer is enough. Also found
a bug in lexing x_0=1;, I think there's a logic error in the
tokenization step.